+++ app/src/main/java/com/nextgis/mobile/fragment/LayersFragment.java
@@ -873,6 +873,13 @@ public class LayersFragment
                 application.sendEvent(GA_LAYER, GA_CREATE, GA_IMPORT);
                 ((MainActivity) getActivity()).addLocalLayer();
                 return true;
+            case R.id.menu_add_underlay_file:
+                ((MainActivity) getActivity()).addUnderlayFile();
+                return true;
+            case R.id.menu_add_underlay_catalog:
+                startActivity(new Intent(getActivity(), com.nextgis.mobile.activity.UnderlayCatalogActivity.class)
+                        .putExtra("choose", true));
+                return true;
             case R.id.menu_add_by_url:
                 application.sendEvent(GA_LAYER, GA_CREATE, GA_NGW);
                 ((MainActivity) getActivity()).addNGWLayerByUrl();